Moving Interfaces and cleaning up files
