+++ python/Ganga/GPIDev/Adapters/ICredentialInfo.py
