+++ python/Ganga/GPIDev/Adapters/ICredentialRequirement.py
