+++ python/Ganga/GPIDev/Credentials/CredentialStore.py
@@ -9,8 +9,8 @@ from Ganga.GPIDev.Base.Proxy import export
 from Ganga.GPIDev.Schema import Schema, Version
 
 from Ganga.Core.exceptions import CredentialsError, GangaKeyError
-from .ICredentialRequirement import ICredentialRequirement
-from .ICredentialInfo import ICredentialInfo
+from Ganga.GPIDev.Adapters.ICredentialRequirement import ICredentialRequirement
+from Ganga.GPIDev.Adapters.ICredentialInfo import ICredentialInfo
 
 logger = Ganga.Utility.logging.getLogger()
